w3c-reports/12092014/gtk/runned-tests/html/semantics/forms/the-input-element/text-expected.txt

Text input element


PASS Value returns the current value for text 
PASS Setting value changes the current value for text 
PASS Value sanitization algorithm should strip line breaks for text 
PASS files attribute must return null for text 
PASS valueAsDate attribute must return null for text 
PASS valueAsNumber attribute must return NaN for text 
FAIL list attribute must return null for text assert_equals: expected (object) null but got (undefined) undefined(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:330:19
http://localhost:8800/html/semantics/forms/the-input-element/text.html:86:24
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
global code@http://localhost:8800/html/semantics/forms/the-input-element/text.html:84:11)
PASS stepDown does not apply for text 
PASS stepUp does not apply for text 
PASS Value returns the current value for search 
PASS Setting value changes the current value for search 
PASS Value sanitization algorithm should strip line breaks for search 
PASS files attribute must return null for search 
PASS valueAsDate attribute must return null for search 
PASS valueAsNumber attribute must return NaN for search 
FAIL list attribute must return null for search assert_equals: expected (object) null but got (undefined) undefined(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:330:19
http://localhost:8800/html/semantics/forms/the-input-element/text.html:86:24
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
global code@http://localhost:8800/html/semantics/forms/the-input-element/text.html:84:11)
PASS stepDown does not apply for search 
PASS stepUp does not apply for search 
